// src/hooks/api/carsInRadius/carsInRadius.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Scalar for receiving email from client */
  Email: any;
};
// ...
export type BusinessType = {
  __typename?: 'BusinessType';
  address: Scalars['String'];
  apiKey: Scalars['String'];
  brandName: Scalars['String'];
  companyNumber: Scalars['String'];
  displayName: Scalars['String'];
  emailAddress: Scalars['String'];
  id: Scalars['String'];
  isFleetInsuranceInPlace: Scalars['Boolean'];
  locationName: Scalars['String'];
  logoImage?: Maybe<ImageType>;
  name: Scalars['String'];
  phoneNumber: Scalars['String'];
  reviews: Array<ReviewType>;
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
  tradingHours: Scalars['String'];
  type: BusinessTypeEnum;
  users: Array<User>;
};

export enum BusinessTypeEnum {
  BundlePartner = 'BundlePartner',
  Insurer = 'Insurer',
  Reseller = 'Reseller'
}
// ...
export type CarFakeLocationType = {
  __typename?: 'CarFakeLocationType';
  lat: Scalars['Float'];
  lon: Scalars['Float'];
  radius: Scalars['Int'];
};
// ...
export type CarSearchResultCarRowType = {
  __typename?: 'CarSearchResultCarRowType';
  addressPublic?: Maybe<Scalars['String']>;
  available?: Maybe<Scalars['Boolean']>;
  business?: Maybe<BusinessType>;
  id: Scalars['String'];
  imagesUrls: Array<Scalars['String']>;
  isAvailableToBuy?: Maybe<Scalars['Boolean']>;
  location: CarFakeLocationType;
  mainImageUrl?: Maybe<Scalars['String']>;
  makerName: Scalars['String'];
  modelName: Scalars['String'];
  pricePerDay: Scalars['Int'];
  pricePerHour: Scalars['Int'];
  pricePerWeek: Scalars['Int'];
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
};

export type CarSearchResultObjectType = {
  __typename?: 'CarSearchResultObjectType';
  cars: Array<CarSearchResultCarRowType>;
  overallCount: Scalars['Int'];
};

export enum CarTransmissionTypeEnum {
  Automatic = 'Automatic',
  Manual = 'Manual'
}

export type CarType = {
  __typename?: 'CarType';
  addressPrivate?: Maybe<Scalars['String']>;
  addressPublic?: Maybe<Scalars['String']>;
  availabilityPeriods: Array<CarAvailabilityPeriodType>;
  availabilityPeriodsComputed: Array<CarAvailabilityPeriodComputedType>;
  bookedPeriods: Array<CarAvailabilityPeriodType>;
  business?: Maybe<BusinessType>;
  createdAt: Scalars['Float'];
  details: CarDetailsType;
  detailsRequested?: Maybe<CarDetailsType>;
  fakeLocation?: Maybe<CarFakeLocationType>;
  features: Array<CarFeatureType>;
  flags: CarFlagsType;
  id: Scalars['String'];
  isAvailableToBuy: Scalars['Boolean'];
  isDraft: Scalars['Boolean'];
  pricePerDay?: Maybe<Scalars['Int']>;
  pricePerHour?: Maybe<Scalars['Int']>;
  pricePerWeek?: Maybe<Scalars['Int']>;
  realLocation?: Maybe<CarRealLocationType>;
  reviews: Array<ReviewType>;
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
  salesInfo?: Maybe<CarSalesInfoType>;
  user?: Maybe<User>;
  visible: Scalars['Boolean'];
};
// ...
export type ImageType = {
  __typename?: 'ImageType';
  id: Scalars['String'];
  url: Scalars['String'];
};
// ...
export type Query = {
// ...
  cars: Array<CarType>;
  /** Returns cars in given radius filtered by parameters */
  carsInRadius: CarSearchResultObjectType;
// ...
};
// ...
export type ReviewType = {
  __typename?: 'ReviewType';
  business?: Maybe<BusinessType>;
  id: Scalars['String'];
  note: Scalars['String'];
  resourceId: Scalars['String'];
  reviewType: ReviewTypeEnum;
  score: Scalars['Int'];
  user?: Maybe<User>;
};
// ...
export type User = {
  __typename?: 'User';
  bannedAt?: Maybe<Scalars['Float']>;
  business?: Maybe<BusinessType>;
  businessUserRole?: Maybe<BusinessUserRoleEnum>;
  callRequest?: Maybe<CallRosterRequestType>;
  communicationPreferences?: Maybe<CommunicationPreferencesType>;
  createdAt: Scalars['Float'];
  details: UserDetailsType;
  detailsRequested?: Maybe<UserDetailsType>;
  email: Scalars['String'];
  id: Scalars['String'];
  internalDetails: UserInternalDetailsType;
  invites: Array<BundleSubscriptionResellerDiscountType>;
  isAdmin: Scalars['Boolean'];
  isBusinessInvitationPopupClosed: Scalars['Boolean'];
  isOwnerProfileSetupFlowInitiated: Scalars['Boolean'];
  /** @deprecated please use communicationPreferences instead */
  marketingPermissionsProvided: Scalars['Boolean'];
  paymentMethods: Array<PaymentMethodType>;
  reviews: Array<ReviewType>;
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
  stripeConnectedAccountId?: Maybe<Scalars['String']>;
  stripeConnectedAccountSetupRequired?: Maybe<Scalars['Boolean']>;
  stripeConnectedAccountUrl?: Maybe<Scalars['String']>;
  stripeCustomerId?: Maybe<Scalars['String']>;
  stripeCustomerUrl?: Maybe<Scalars['String']>;
  subscriptions: Array<BundleSubscriptionType>;
  tags: Array<TagType>;
  validAsCarOwner: Scalars['Boolean'];
  validAsCarRenter: Scalars['Boolean'];
};
// ...
export type CarsInRadiusQuery = { __typename?: 'Query', carsInRadius: { __typename?: 'CarSearchResultObjectType', overallCount: number, cars: Array<{ __typename?: 'CarSearchResultCarRowType', id: string, makerName: string, modelName: string, pricePerDay: number, reviewsAverageScore?: number | null, reviewsCount: number, available?: boolean | null, addressPublic?: string | null, mainImageUrl?: string | null, business?: { __typename?: 'BusinessType', logoImage?: { __typename?: 'ImageType', id: string, url: string } | null } | null, location: { __typename?: 'CarFakeLocationType', radius: number, lat: number, lon: number } }> } };